*: improve godoc
Cleans up the API surface and make sure to document all public members.

Note: This commit removes these previously-exported members:
- Taskfile
- TaskfileTask
- Func
